Implement client-side disk type auto discovery
In server create, the API does not automatically select a correct root
device to install the OS on. The OS disk parameter is required. And
since the client cannot inherently know the OS disk device name, we
can query the core endpoint to check the disk_type in the core metadata.

If the disk is a NVMe type, we use /dev/nvme0n1 as the OS disk, otherwise
we use /dev/sda (which was the previous default.)